Config manager added to core components, remade defaults
