+++ src/main/scala/org/fs/mael/ui/prefs/GlobalPreferences.scala
@@ -7,9 +7,8 @@ import org.eclipse.jface.preference.PreferenceDialog
 import org.eclipse.jface.preference.PreferenceManager
 import org.eclipse.jface.preference.PreferenceNode
 import org.eclipse.swt.widgets.Shell
+import org.fs.mael.core.config.ConfigManager
 import org.fs.mael.core.utils.CoreUtils._
-import org.fs.mael.ui.ConfigManager
-import org.fs.mael.ui.ConfigOptions
 
 class GlobalPreferences(val configMgr: ConfigManager) {
   import GlobalPreferences._
@@ -30,21 +29,48 @@ class GlobalPreferences(val configMgr: ConfigManager) {
 }
 
 object GlobalPreferences {
+  import org.fs.mael.core.config.ConfigOption._
+
+  val DownloadPath: SimpleConfigOption[String] = SimpleConfigOption("main.downloadPath")
+  val NetworkTimeout: SimpleConfigOption[Int] = SimpleConfigOption("main.networkTimeoutMs")
+  val SortColumn: SimpleConfigOption[String] = SimpleConfigOption("view.sortColumn")
+  val SortAsc: SimpleConfigOption[Boolean] = SimpleConfigOption("view.sortAsc")
+  val ActionOnWindowClose: RadioConfigOption[OnWindowClose] =
+    new RadioConfigOption("main.actionOnWindowClose", OnWindowClose.values)
+  val MinimizeToTrayBehaviour: RadioConfigOption[MinimizeToTray] =
+    new RadioConfigOption("main.minimizeToTrayBehaviour", MinimizeToTray.values)
+
+  sealed abstract class OnWindowClose(id: String, prettyName: String) extends RadioValue(id, prettyName)
+  object OnWindowClose {
+    object Undefined extends OnWindowClose("", "Prompt")
+    object Close extends OnWindowClose("CLOSE", "Close")
+    object Minimize extends OnWindowClose("MINIMIZE", "Minimize")
+    val values = Seq(Undefined, Close, Minimize)
+  }
+
+  sealed abstract class MinimizeToTray(id: String, prettyName: String) extends RadioValue(id, prettyName)
+  object MinimizeToTray {
+    object Never extends MinimizeToTray("NEVER", "Never")
+    object OnClose extends MinimizeToTray("ON_CLOSE", "On window close")
+    object Always extends MinimizeToTray("ALWAYS", "Always")
+    val values = Seq(Never, OnClose, Always)
+  }
+
   class MainPage extends RichFieldEditorPreferencePage(FieldEditorPreferencePage.FLAT) {
     def createFieldEditors(): Unit = {
-      row(ConfigOptions.DownloadPath) { (option, parent) =>
+      row(DownloadPath) { (option, parent) =>
         new DirectoryFieldEditor(option.id, "Download path:", parent)
       }
 
-      row(ConfigOptions.NetworkTimeout) { (option, parent) =>
+      row(NetworkTimeout) { (option, parent) =>
         new IntegerFieldEditor(option.id, "Network timeout (ms, 0 means no timeout):", parent).withCode { field =>
           field.setValidRange(0, 7 * 24 * 60 * 60 * 1000) // Up to one week
         }
       }
 
-      radioRow("Action on window close:", ConfigOptions.ActionOnWindowClose)
+      radioRow("Action on window close:", ActionOnWindowClose)
 
-      radioRow("Minimize to tray:", ConfigOptions.MinimizeToTrayBehaviour)
+      radioRow("Minimize to tray:", MinimizeToTrayBehaviour)
     }
   }
 }
